SearchPoint: remove DebugFlag, manage "projected" manually
It turns out that DebugFlag is not zero-overhead in the optimised
build: the C++ standard mandates that an empty struct must have a
non-zero size.  There have been ugly inheritance hacks to work around
this (look at how allocators instances are managed in most STL
implementations), but that's even more ugly than just managing the
flag manually with #ifdefs.