GlideResult: use values with wind drift in Add()
The Add() method and most of GlideResult is still in a sad state:
sometimes, the results include circling wind drift, and sometimes,
they do not.  I added pure_glide_height to resolve some of the
confusion, but in reality, I added more, because methods like
GetRequiredAltitude() suddenly didn't use wind drift anymore - which
seems logical, but it breaks the assumptions in Add().

This patch adds another set of helper methods with wind drift, and
switches Add() to use them.