Form/List: don't wrap focus to next control on Altair
This fixes a usability disaster: when the last item is selected, and
the user turns the wheel further, XCSoar will suddenly switch to the
next dialog page, because this wheel has two meanings, depending on
the context.  This adjusts the list control to use the same meaning as
the edit control.