GlideSolvers/MacCready: fix pure_glide_height in partial solutions
In function SolveGlide, pure_glide_height was calculated for the whole task
distance even when a partial solution was allowed. This resulted in a wrong
pure_glide_height when the glide solution calculated by MacCready::Solve is
the addition of a partial climb-cruise and partial glide.

SolveGlide is by definition the solution for a task with no climb component,
so in this case pure_glide_height is equal to height_glide.