CommandLine: Silently ignore anything not understood on WIN32 targets
Now xcsoar only exits if an positively identified option is specified
with incorrect parameters.

Temporary workaround until a robust solution is found. We really need
some way to communicate the usage message to users on WIN32 targets.

Solves ticket #1772 which was probably caused by ' ' characters in
the path leading to xcsoar (the first parameter was
"Laptop\Downloads\XCSoar" without the " characters) despite
comment #6.