Driver/FlyNet: Added low-pass filter to the raw pressure input
This was suggested by John Wharington and calms the input value.
Since the resolution of the pressure sensor is only 1 Pa, a small change
could cause a big spike in the vario data. Due to the low-pass filter
we don't have any sudden jumps of a whole Pa in the output anymore.