Hardware/ModelType: add model "LX MiniMap"
Unfortunately, the LX MiniMap cannot be auto-detected; it's a standard
PNA with Windows CE Core and no special registry settings.  So the
user has to configure it explicitly.

This patch adds a key map to GDI/Transcode.cpp which translates a few
of the buttons to VK_APPx (see TRAC #1774).