RoutePolar: assume only straight glide in intersection()
This fixes a major breakage in the terrain warning calculation.

Commit c730a601f changed the min_height calculation, and caused the
regression; a bogus terrain warning would always be calculated right
before the destination, given that the destination is not reachable
according to GlideResult.

The commit however is correct, previous behaviour happened to be two
bugs compensating each others.  With a large (bugged) min_height
value, h_diff is always zero, and an intersection is usually not found
in "gradual climb to destination" mode (unless the aircraft is already
below the next mountain top).  This suppressed the bogus terrain
warning.

With the corrected min_height value and an unreachable destination,
intersection() assumes cruise-climb.  The cruise-climb mode however is
severely bugged, its formula will always result in a terrain collision
in the final "glide" part.

This patch is a radical approach at fixing the problem: it removes
nearly all of the code from intersection(), leaving only the pure
glide mode.  This is the only thing we should ever need for the
terrain warning.  Assuming cruise is wrong, because this should be a
realistic calculation based on current altitude; when you assume that
you can always get lift, then your potential glide path is infinite.
While this may be desirable for a pilot, it is not realistic.