NMEA/Derived: Don't expire calculated.wind_available
The last update time of different wind sources was simply copied to
calculated.wind_available. However the expiry threshold for
calculated.wind_available was considerably shorter than for the real
wind sources often causing the immediate expiry of it here leading to
constant available/!available transitions.

Since DerivedInfo::Expire() and WindComputer::Select() run at different
times this lead to inconsistent availability information in different
areas of xcsoar. E.g. for the task engine the wind was invalid but for
infoboxes and the wind arrow it was valid.

Fixes trac issue #1807.