Android/InternalGPS: fix deadlock when internal GPS is disabled
The synchronous updates caused a severe deadlock (TRAC #1816): the
(native) UI thread creates the InternalGPS object, which posts a
message to the Java main thread and waits for its completion.
Meanwhile, the Android Activity gets paused, and the Java thread
waits for the native UI thread to handle the PAUSE pause event.

This commit makes update() and setLocationProvider() non-blocking.
The new method close() waits for completion.

It implements a new way of shutting down the object; a new flag tells
us that the native code has requested to dispose the InternalGPS
object, and with this flag set, native methods will not be called.  A
reference counter keeps track of how many native method calls are
currently running.  The method close() will wait for this reference
counter to become zero.