GlideComputerAirData: Removed duplicate fabs()
The inner fabs() was useless and led to potential bugs in corner cases
where g_load is between -0.75 and -1.25. If the g_load is in that
range we shouldn't use the data because of the accelerated state that
leads to untrustworthy data. Only if the g_load is between 0.75 and 1.25
the data should be used.