InfoBoxes: add NO_HORIZON compile option
This disables the artificial horizon InfoBox, while
leaving the indexes of the InfoBoxes the same,
so switching from between the crippled version and
the full version will not mess up InfoBox configurations.
It displays "Horizon: Disabled" on the splash screen and
also shows "Disable" in the InfoBox Comment.
It requires no new additional strings for the translation team.