Engine/Tasks/AbortTask: Fix WP selection
Previously a once selected alternate WP (the selection process is
done continuously even if the abort task is not active) was retained
until it was not anymore in the list of possible alternates. This lead
to the confusing situation that on activating the abort task this far
away and usually not reachable WP is preselected as target, e.g. the
HOME waypoint (because it was the nearest one at takeoff) is kept
preselected until it is not anymore in the area examined for possible
alternates (a radius of 50-100km around the current position).

This commit forces the selection of the best alternate continuously
until the abort task is activated. In an activated abort task it then stays
selected until it drops out of the area considered or the user selects
a different one.

Fixes trac issue #1907.