GlideComputerRoute: limit task distance to a reasonable value
GeoVector::EndPoint() may crash with an assertion failures because the
formula in FindLatitudeLongitude() is not suitable for large longitude
differences (TRAC #1969).  This patch adds a GeoVector distance limit
of 200km.  This is reasonably large for real world flying, and avoids
the bug (unless you're flying at one of the poles).