Makefile: revert "ar" fallback
This reverts a part of commit eebc4a ("Configure: set local toolchain
path and a toolchain binary extension").  The part which adds a
fallback for "ar" doesn't work at all, and broke every cross build.
It always falls back, because the expression does not check for
available executables, it only checks if "ar" is a file in the current
directory.