WaypointRenderer: fall back to MacCready if reach fan is empty
The WaypointRenderer uses the RoutePlanner's ReachFan to query the
arrival heights.  This fails to work when there is no terrain data,
which means no arrival heights can be calculated (TRAC #2018).

So this patch falls back to the MacCready library to calculate the
arrival height, instead of vainly querying the route planner again and
again.