GlueMapWindow: Changed visibility of OffsetHistory member functions
Previously all members were either private or protected, including the
constructor. Access was only allowed for the GlueMapWindow class because it
was declared as a friend class.

This patch lifts these restrictions by making the member functions public and
removing the friend class which decreases the attribute visibilities to
private for the GlueMapWindow class.