Math/fixed: reduce risk of overflow in fixed-point hypot()
In the ETE/ETA calculations on big tasks, calls to hypot() could
overflow due to unsafe use of sqr(), leading to wrong results (TRAC
#2066).  This commit reduces the risk by discarding the lower 3 bits
of both input variables before calling sqr().  It does not solve the
problem, it merely moves it to an input range that is not required by
our formulas.  It adds an overhead of 3 bit shift operations for each
hypot() call.

Credits to Andreas Pfaller for debugging this issue.