CAI302: track mode in CAI302Device::EnableNMEA
This fixes the problem where the CAI302 disconnected for 30 seconds after
the user updates the MC in XCSoar.

This solution tracks the NMEA mode in the CAI302 driver, and stops the RxThread if
a mode other than NMEA mode is present when EnableNMEA() is called.  This logic
is consistent with the CAI302Device::CommandMode(), CAI302Device:: UploadMode()
and CAI302Device::DownloadMode() functions.

Cause: the CAI302::PutMacCready does not stop the RxThread.  However
DeviceDescriptor::PutMacCready uses ScopeReturnDevice which calls CAI302::EnableNMEA()
and WaitCommandPrompt which fails if RxThread is running.
As a result DeviceDescriptor::PutMacCready was incorrectly leaving the 302 device in
command mode when EnableNMEA() called CommandMode() and WaitCommandPrompt
failed.