LocalPath.cpp: use external_sd only if it's not mounted as tmpfs.
This aims to fix the problem on some Android 4 devices which have
a /sdcard/external_sd directory where an external sd would be mounted.
If there's no external sd card inserted, the external_sd directory
still get's mounted as tmpfs by android.

This checks if there's an entry in /proc/mounts for the external_sd
which is not mounted as tmpfs and if so, it's used for XCSoarData storage.

Might fix ticket #2051