Android: save logcat after a crash
It's hard for users to understand how extracting the logcat works, and
more often than not, your phone doesn't have any program capable of
saving the logcat during the flight.  This simple patch invokes the
Android "logcat" utility during startup.  If a crash was found, the
log file is saved in the new "crash" directory.  Hopefully, this will
lead to better bug reports.  This requires the Android permission
READ_LOGS which must be confirmed by the user on installation /
upgrade.