OrderedTaskBehaviour: move attributes to TaskFactoryConstraints
Move those attributes that are constant at compile time (per factory
class) to the new struct TaskFactoryConstraints.  That will provide a
better separation of compile-time constants and user settings.

As a side effect, this removes a few attributes from the XML task
files.  But these weren't useful anyway, because they could not be
edited by the user, and were discarded after loading from the XML
file.