QuadTree: Don't change square_distance to max_distance() in FindNearestIf()
If Bucket::FindNearestIf() did not find a suitable Leaf it returns
std::pair<const_iterator(), max_distance()>. If we would assign this
to square_range we would open the following searches to maximum range.
We only assign a new opimized square_range if a Leaf was found and with an
assert() we check if that one is now really below the original
square_range.