Specify 'glide ratio' as 'glide ratio over ground'/'lift/drag ratio'
'Glide ratio' is a non-specific term which can refer to the ratio
of horizontal to vertical motion with reference to either the airmass
or the ground.

Usage of ground-referenced glide ratios are referred to by the term
'glide ratio over ground', or 'glide ratio' / 'GR' for short.

Air-referenced glide ratios are specified as 'lift/drag ratio' /
'L/D ratio' / 'LD', which is equal to the air-referenced glide ratio
when flying at constant speed.