CalculationThread: implement forced recalculation
When the task is edited or a similar event occurs, the new function
ForceCalculation() forces a recalculation of the task statistics, even
if no GPS fix is available.  This fixes an ancient XCSoar quirk: the
"target" button is not available until XCSoar has a GPS fix.  This
could be related to TRAC #2183.