DisplayMode: new option to disable the "final glide" display mode
The need for a "final glide" display mode is a matter of taste.  Some
find it useful, but others find it plain annoying to be lost with an
unconfigured set of InfoBoxes at the end of a flight.  This patch
adds a knob to enable/disable this display mode, disabled by default
for new users, enabled for existing users.

I'm not yet sure if the InfoBoxesConfigPanel is the best place to put
this option.