TaskListPanel: add "More" button to show all task types
The "More" button replaces the old "Back" button which was not so
useful, because the same could be done with clicking on the tab bar,
and other panels don't have a "Back" button either.

This has two effects: first, the list is not so cluttered.  Second,
enumerating tasks of *.cup and *.igc is very expensive, because the
file contents have to be read and parsed, and therefore we're
defaulting to "false".