Device/List: allow up to 6 devices
With IOIO, Bluetooth and internal GPS, one can easily exceed the old
limit of 4.  This is a trivial patch which adds two more devices for
Linux and Android.  We could add even more, but that wouldn't add any
real value currently.  WIN32 is still limited to 4, because it is
unlikely that a Windows user will have more than that.