IOThread: obey iterator limits in Update()
When erasing the first element, the iterator could be decremented to
before the first element, i.e. before begin().  That made valgrind
cry.

Instead of decrementing the iterator, the increment is now moved out
of the final segment of for() and incremented as needed.  This fixes a
potential crash in I/O thread.