Device/Descriptor: wait until Port gets ready
Before calling Device::EnableNMEA(), we must ensure that the Port is
ready, for example that an incoming TCP connection was established or
an IOIO board was found.

This means that the OpenDeviceJob will sometimes never finish for TCP
ports, but that's not a problem, because this thread can easily be
cancelled on demand.