Screen: add EGL backend (without SDL) - experimental!
This will allow using XCSoar on Linux without SDL.  Most prominent
hardware for this is the Raspbarry Pi.  Beware, this is unfinished
work in progress.  The event queue does not process input events yet,
only timers work so far.  Terrain is broken, and on VideoCore (Pi),
nearly everything is broken.