InfoBoxWindow: open dialog on short click (if there is one)
Without hardware keys, it's impossible to manipulate a focused
InfoBox.  InfoBoxes that have a dialog shall open the dialog after a
single short click, instead of requiring the user to hold the finger
down for at least one second.  This is a short-cut for those InfoBoxes
that already have a dialog, and the remaining ones should be extended
soon.