SampledTaskPoint: always use the OZ boundary
GetSearchPoints() always returns the OZ boundary, no matter if
IsBoundaryScored() is true or false.  All turn points should be
optimised for minimum/maximum distance.  This was the wrong way round
previously, and this patch fixes a bug, not a missing feature.  This
happened to have worked correctly for AAT, because AAT overrides the
minimum distance search point anyway.

As a side effect of this bug fix, we need to set FinishPoint's
boundary_scored value to false, to retain optimised finish points.