AirspaceParser: Fixed airspace arc closing bug
When an arc was defined with start and end point very close together, but
with a rotation direction so that a circle with a small piece missing is
expected, the parser just connected the two points directly due to wrong
threshold handling.

Now we make sure that the end angle is always on the correct side of the
start angle to prevent this problem.

This might also slightly speed up the functions since the AsBearing() call
is removed from the loop.

This fixes bug #2360