NMEA/Derived: add attribute "ordered_task_stats"
Some libraries explicitly want data of the OrderedTask, even if
another task is active.  We used to solve this by copying a few
TaskStats attributes to CommonStats, but this solution is better, and
we can remove these duplicate attributes now.  A side effect of this
patch is that a few InfoBoxes and dialogs are now able to work even in
abort or goto mode.