Revert "Form/Frame: call Window::SetFont()"
This reverts commit 5f45451c311a8f020bd505f70d8d7bdd84c682c1.

WM_SETFONT is documented to work only with "controls", see:

 http://msdn.microsoft.com/en-us/library/windows/desktop/ms632642.aspx

The web site does not define what a "control" is, but practice shows
that this means dialog controls ("TEXT", "EDIT", ...) and that the
message is ignored by the default window procedure.  Therefore, we
need to select the HFONT explicitly for each WM_PAINT.

TRAC #2507 has now been solved by not using WndFrame for font
previews.