Look: don't use bold font for list items
A bold font is less readable than a medium font.  Therefore, bold
should only be used to emphasise small portions of the screen.  List
items however are normal data, they do not need to be emphasised.
When a list is being displayed, its text will cover most of the
screen, and emphasising everything is pretty pointless, even more so
if you choose the emphasis so it makes the text less readable.