Look/Fonts: use non-bold font for InfoBox title/comments
The InfoBox comment should be de-emphasised, because its data is
useful, but needs to be read by the user only rarely.  The title is a
bit more important, but does not deserve emphasis.  After all, the
value is most important in the InfoBox, and the only thing that will
be read by the user frequently.