GaugeVario: Fixed typo
Apparently this typo has been there from the start of the OpenGL
implementation and nobody noticed it.

The bugs display needs to be redrawn if either the bugs value has changed or
the canvas is NOT persistent. With the previous behavior the bugs were only
drawn if they changed on backends with a non-persistent canvas.

See commit 4965c22952dab33d30bd7af8dc16eeb24c5f4f73