Formatter/TimeFormatter: Indicate usage of mm:ss format for short timeframes
From a value like "05:03" it is not immediately obvious if this means
"5 hours + 3 minutes" or "5 minutes + 3 seconds". The only visual indication
in xcsoar's infoboxes is the appearance of the seconds in the infobox
comment field if the former is shown. However for short time-frames
the missing second is easily missed. Try to improve the situation
by separating minutes and seconds with "'" for short time-frames.

Signed-off-by: Andreas Pfaller <pfaller@gmail.com>