Math/Angle: don't use optimised multiplication for Angle*fixed
Reverts the second part of e4b9b576.  Turns out that the
CatmullRomInterpolator does care a lot about the bits discarded by the
fast_mult() call, because it passes very small values (TRAC #2537).
The unfortunate bottom line is we can't predict the dynamics of the
two factors, and we must not discard any bit in the default operator
without knowing from where this was called.