Look/Fonts: Revert the recent font changes for topography labels
This reverts commit 964c3a7485e5a59f291807e090474b826e9c67c8:

An UI regression. Previously topography labels were immediately
obvious simply by looking at the label because of the distinct
rendering in italic. Now you have to scan the surrounding area
to notice the different font size (if it even is sufficiently
different). Even if italic is harder to read (a problem I don't
notice for the labels) this additional scan time will invalidate
the assumed advantages.