AirspaceRenderer: Enable the explicit use of airspace paddings on PPC200x
If the user explicitly chooses to trade speed for better looks we should
not force him to use the fill-all rendering style. The default on ancient
hardware should obviously be optimized for speed, but the user should have
the option to change it. Giving him the option to change it but not really
changing it as previous versions did is certainly a bad idea.