UIUtil/KineticManager: Fix jumps upon mouse up
If one tried to accurately position the map (or other UI element using the
KineticManager) by dragging, slowing the drag and releasing the mouse
the map often jumped by dozens of pixels since during the final steady
period (no movement) the velocitiy was not updated (no mouse move events).
This commit uses a 200 ms threshold with no movement to detect the attempt
of accurate positioning.