MapWindow/GlueMapWindowEvents: Disable MapItemList for long mouse clicks
It is often desirable to accurately position the map to get a precise
reading of elevation and position (upper right hand corner info).
However this often has the side-effect of popping up the MapItemList
on mouse up since the MapItemList was only disabled if the map was
panned a considerable distance. Add an additional time based criteria
for disabling the MapItemList trigger.