Volkslog:Implement first_char timeout, let readdir return data_length
When the command to download the the main flightlog
data is send to the VL it starts calculating security.
Therefore the reply to the first ACK takes very long.
For this reason a special timeout for the first char
had to be implemented in the ReadBulk() routing and
subsequently in the routines SendCommandReadBulk() as
well.
Also let readdir() return the length of the read data
so it can be passed to conv_dir to fix the parsing
bug there.