vlapi: SendCommandReadBulk() implement optional longer timeout for first char
When downloading the binary log from the logger it is necessary to
wait a longer period after the first ACK is sent to the logger until
the first char is received.
This is because the logger starts calculating file security before the
first byte is sent. This may take several minutes.
The additional parameter needs to be implemented in ReadBulk() and
subsequently also in SendCommandReadBulk()