vlconv: fix conv_dir()
Testing has shown, that the function conv_dir() as it is does
not contain a valid condition to end the parsing loop of data
in the buffer.
Reasons might be a change in the api in the past, or (less likely)
the Volkslogger used for testing having a different protocol than
the api expects.
Some hints in the code point to a not finished work in the past
which left the code not working.
The exact details of the volkslogger protocol are not known,
so it is not possible to tell why this is the case.
The working solution to the problem is to pass the number of chars
in the buffer to the function and stop parsing after all data was
processed.