Driver/FLARM: remove bogus SetProgressPosition() calls
The SetProgressRange(10) call was using a random value, and
SetProgressPosition() is being called with more random values.  This
is bogus information, because the "progress" bar will indicate that
the operation is finished, only to start from the beginning.  It's an
abuse of the progress bar to generate an animation that doesn't say
anything useful.