Renderer/TrackLineRenderer: Don't check basic.attitude.heading_available
This fixes trac issue #2616.

For the internal "fallback" calculation of the heading heading_available
is never set to true. It is also not possible to set this to true since in
that case the recalculation would be skipped until it expires.

Since we are guaranteed to calculate a heading if basic.track_available
is true that check should be sufficient.