vlapi:Replace new/delete by std::vector
The line with delete[] caused a segmentation fault.
The new/delete[] was replaced by switching to
std::vector.
This allows us to do some more cleanup:
-DIRECTORY structure obsolete.
-Since we have conv_dir() now using std::vector.
The first "count only" call of the function is
no longer necessary.
The original use was to determine the count of
entries so the dynamic array could be initialized
with the correct size. std::vector makes that
obsolete.
-int number_of_flights used in conv_dir() is
no longer necessary