Form: focus the first client_area control
Fixes a problem on Windows CE: the dialog itself does not have the
WS_EX_CONTROLPARENT style, and therefore is not supposed to have
controls.  Windows (on the desktop) ignores this problem, but
GetNextDlgTabItem() on WinCE will fail and will return the dialog
handle itself.  Focusing it will put the keyboard focus nowhere, and
will disable all harware keys.  This is pretty fatal on the Altair.