Android/Bluetooth: join the connect thread in close()
Fixes a race condition leading to a crash (TRAC #2636, #2656): when
the port is being closed just when BluetoothClientPort is in the
middle of socket.connect() and setPort(), a new InputThread gets
created that is orphaned instantly, nobody is aware it exists (and
refers to the NativeInputListener).  After close() returns, the native
code deletes the NativeInputListener.  When the orphan InputThread
invokes the NativeInputListener, XCSoar crashes.

The solutionb is to wait until BluetoothClientPort's thread finishes
before cleaning up the ProxyAndroidPort.  This ensures that no such
orphan objects exist.