NMEA/Attitude: Provide possibility for checking if heading is usable
For the internal "fallback" calculation of the heading heading_available
is never set to true. It is also not possible to set this to true since in
that case the recalculation would be skipped until it expires.

Provide an additional flag to allow checking the "usability" of the
heading regardless of how it was obtained.