Volkslogger Declaration: Keep NavDatabase
Up to now the VL_Declaration wiped out the NAV-Database(Turnpoints)
on the Logger. This is because the VL can only receive declaration
and Waypoints in one block.
This patch extends the functionality of VLAPI to read the raw
database from the the logger, update only the relevant section to
change the declaration, and finally writes the updated raw data back.
The functions read/write_db_and_declaration() have both issues and
therefore were not used. This raw data approach is imo a lot safer
and a little faster since there has be no converting of the database
back and forth.