Altair RemoteStick Patch
This Patch handles all necessary changes to use the Altair Remote Stick
with the new XCSoar Versions.

Mainly:
     Key-Definitions have been made in order to use the Key-Events
(created by the RemoteStick) with the Event-Handler.
     Additional handling of Key-Events in the Quickmenu in order to
control the Menu by the RemoteStick (changing focus of button, "click"
on button, Escape-Function to leave the Menu again)
     Additional Keyevents in the Analyis-Page (to change the pages with
the RemoteStick and to leave the Menu)