CommandLine: add parameter "-datapath"
This Patch allows to specifiy the used DataPath for the XCS Data when
starting XCS with command line.
This will allow the direct use of the storage card in the Altair by XCS
(getting rid of the all omipresent storage problem in Altair), when the
AltairSync program will be updated by Triadis (otherwise the call is
still the same).