VarioBar
Implements a "VarioBar" on the right edge of the map similar to the
FinalGlideBar on the left edge. It displays the current vertical speed,
the averaged vertical speed, as well as the current MC setting in a
compact manner. The VarioBar can be enabled/disabled in the
"Gauges->Flarm,Other" panel.