CirclingWind: fix bogus time warp
When the CirclingWind object gets activated an the last thermal was
more than 5 minutes ago, CirclingWind::NewSample() detects a time warp
because last_*_available was never cleared.

Previously, only the second thermal was used after a long glide.  This
is bad, because the first thermal is important to know the new wind
for the next glide.  On a typical flight, this increases the number of
wind measurements by more than 50%, marking a major improvement in the
algorithm.