Device/Descriptor: implement port dumping
A new button in the device manager enables dumping for a few minutes.
Previously, this feature was only available in debug builds and had to
be configured explicitly and permanently.  This commit avoids the
problem that can occur (log file filling the SD card), while still
allowing users to send us dumps for better bug reports.