Android: import IOIOLib as a git submodule
Previously, to build the Android package, you had to have a copy of
IOIOLib somewhere on your hard drive, and use the make variable
IOIOLIB_DIR to tell XCSoar's Makefile the location of that directory.
That was cumbersome and error prone, because there was no good way to
specify which IOIOLib version XCSoar expects.

Another idea was copying just the IOIOLib into xcsoar.git, but that
was hard to keep up to date, and maintaining a copy of another project
is not so elegant.

The git submodule refers to the commit that is near to the release
V3.30.

The IOIO feature is no longer optional.