LoggerImpl: use GPSState::real only if location_available
If the logger starts and only a vario is available (but no GPS fix),
the G record will be disabled for the whole IGC file, because
GPSState::real is false.  The problem is that GPSState::real is used
even though nobody ever set it to true prior to getting a GPS fix.

This problem was similar to TRAC #2849.  This variant of the problem
is theoretical, it was never observed in reality.