Look/Fonts: add dedicated dialog fonts
Separate the dialog fonts from the map fonts.  For map fonts, only the
physical size counts (readability), but dialog fonts have different
requirements: a certain number of options must fit on the screen, the
items must be large enough for a finger to tap on, etc.  (These rules
are not implemented yet, but now they can easily be.)

Side effect: the "small" font is not italic anymore.

Fixes TRAC tickets #723 and #2806.