TaskAdvanceMonitor: move code from InputEventsTask
This commit establishes a new subsystem: the "monitors".  A monitor
checks new states after each calculation and invokes some user
interface action when a monitored object changes to a specific state.
The first monitor is the TaskAdvanceMonitor, which replaces the event
"TaskTranstion ready".  The new monitor takes care for closing the
widget when it's not needed anymore.