Screen/Custom/RasterCanvas: avoid redundant clip tests
Previously, every time a clip was applied, both p1 and p2 points are re-tested against
clipping borders (taking 4 comparisons for each point).  However, each time
we clip, we are only adjusting clip1, so we only need to re-test that point.
In simplest case of single clip, we save 4 comparisons; in most complex case,
we save 16 comparisons with this optimisation.