Screen/Custom/RasterCanvas: reduce branches in clip test
Most lines to be drawn will have points inside the clipping area, so
we want tests for these points to be the fastest.  Since left/top dimension
is zero, we can just perform a single check for each axis to see if each dimension
is within range, and a second check as to which edge is breached only if it
fails the first test.