Repetitive warning sound for airspace warnings dialog (enhancement #2952)
Adds repetitive warning sound when approaching or inside the prohibited airspace and
airspace warnings dialog is shown.
This is useful if you miss initial warning sound in sometimes noisy glider
cockpit and are close to airspace.
Warning sound is played with increased frequency as you are closer to airspace.
Enabling of this feature is configurable in Airspace configuration page by
checkbox "Repetitive Sound".