Wind/WindEKF.cpp: tweak initial covariance fix #2961
At the initialisation of Kalman filters, the filter needs estimates of the
initial covariance (read as uncertainty) of the hidden states.  Due to
the nonlinear nature of the algorithm, in certain conditions if the
estimates are asserted to be more accurate than they actually are, the
algorithm can become unstable.  This is what was occurring with #2961; lag
introduced by the update order drives the scale factor towards zero or infinity,
with consequent effects on wind estimates.

Increasing the initial uncertainty of scale factor produces stable results
and equivalent wind estimates (that is, it does not appreciably slow down the
convergence).

This must go in the running for the competition for smallest bug fix
(one character) with most obscure reason behind it.