Map: draw projected path when turning
Assumes constant airspeed and bank angle.

If in circling mode and trail drift mode, the path is NOT adjusted for wind
and the path is therefore wind-relative --- this is the path through the air
that the glider will take.

Otherwise, the path is adjusted for wind and is therefore ground-relative ---
this is the path over the ground the aircraft will take.