dlgConfiguration: disable custom fonts on Altair
On the Altair, we have hard-coded bitmap fonts.  These bitmap fonts
were never available in the font editor, and vector fonts look ugly on
the low-res screen.  Customising fonts is just not worth it on the
Altair.  But the real reason to remove this setting now is that we
need more special hard-coded font settings, and integrating those into
the font customisation system would be too much work.