test/DebugReplayNMEA: provide a stable "clock"
NMEAInfo::clock was copied from NMEAInfo::time, but when the latter
was unreliable or suffered from time warps, the clock could be reset
to zero, marking it "invalid" for Validity objects, and no new time
could ever be parsed.

The new approach in class ReplayClock starts with 1 and applies only
offsets from NMEAInfo::time, but never resets.