DeviceBlackboard: don't overwrite replay_data.clock
The replay class is responsible for this.  Overwriting the value
causes a lot of damage, for example on time warps or when there's no
time value yet - in that case, time_available will never be non-zero
because the clock is still zero, and the clock will be zero until
time_available is set.