AirspaceRenderer, ...: don't pass reference to temporary AircraftState
The temporary object will be deallocated soon, but AirspaceVisibility
and others still have references to the object.  That leads to wrong
visibility results due to random altitude values (TRAC #3061).
Instead of passing a temporary AircraftState object, allocate it on
the stack explicitly and pass a reference to this stack object.

This bug had no effect on gcc versions older than 4.8, because gcc
4.8 frees temporary objects earlier.