BigTrafficWidget: auto-restore page layout only if not FLARM page
If auto-closing the FLARM page is configured as soon as no traffic is
received anymore, call PageActions::DeferredRestore() only if the page
we're about to restore is not a FLARM page.  That would mean that we'd
be calling PageActions::DeferredRestore() over and over, but the
TrafficWidget instance will stay, but will never call Update(),
leaving the last visible traffic on the screen forever (TRAC #3030).