Settings/MapDisplayConfigPanel: Remove "Auto Zoom" On/Off setting
The "On" setting for auto zoom is unlikely to survive a minute
of regular use since a normal zoom action will disable it. It is really
a display option which only makes sense to set via shortcuts (e.g.
the display menu or a gesture) when desired as an alternative to
manual zoom levels which are also only available via shortcuts.