Engine/Airspace/AirspaceWarningManger: Prevent creation of temporary warnings
The AirspaceWarningManger used many temporary warning objects which
were immediately discarded when the warning later turned out to be
not relevant causing a constantly incremented serial number.

This had the side effect of causing invalidation of the
cached airspace fill patterns used in the AirspaceRender since
it relied on the serial to trigger cache updates.