build/targets.mk: make kobo binaries static
Without this, executables fail due to missing shared libraries on kobo.

Either the libraries need to be installed on kobo, or binaries made static.

[mk: This reverts commit 3c7cb0b5 which worked well on a Kobo with the
most recent firmware, but fails with old firmware versions.  More
testing needs to be put into this, to ensure that we can deal with all
firmware versions.]