Revert "Makefile, doc: require Android NDK r9b"
This reverts commit 3405a475f214b81798c35fac5b7e150270baa0d6.

Upgrading to this new NDK version caused severe problems with the
KDTree library.  The call to std::nth_element() crashed due to buffer
overflow.  At this time, I'm not sure if this bug is in the NDK or in
the KDTree library.  Just to be safe on the stable branch, I hereby
reverting the upgrade.