TaskManager: Show task file name of loaded/saved tasks in dialog title
Any modification of the loaded/saved state will display "Default.tsk".
Currently the name will not survive a restart of XCSoar.

In the current version saving of an (unmodified) task causes a full
Commit() to the task engine (like a modified task). This could be
optimized.

Implements trac enhancement request #1924.